Move language implementation into `lib.rs` file
